videoplayer 디자인
videoplayer 디자인

+++ controllers/videoController.js
@@ -42,8 +42,12 @@ export const postUpload = async(req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
+
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -53,7 +57,8 @@ export const videoDetail = async(req, res) => {
         params: { id }
     } = req;
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate("creator");
+        console.log(video);
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -68,7 +73,11 @@ export const getEditVideo = async(req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -94,8 +103,12 @@ export const deleteVideo = async(req, res) => {
     } = req;
 
     try {
-        await Video.findOneAndRemove({ _id: id });
-        res.redirect(routes.home);
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
     } catch (error) {
         console.log(error);
